+++ buildings/templates/buildings/index.html
@@ -259,7 +259,7 @@
             {% endfor %}
           </ul>
 
-          <div class="pagination mt-3">
+          <div class="pagination mt-3 ms-2">
             <span class="step-links">
               {% if user_votes_page.has_previous %}
               <a href='#' hx-swap="outerHTML" hx-target='#activity-tab-content'
