+++ buildings/templates/buildings/survey_rendering_full.html
@@ -370,8 +370,8 @@
   }(HTMLCanvasElement.prototype.getContext);
 </script>
 <script defer src="{% static 'scripts/html2canvas.min.js' %}"></script>
-<script defer src="{% static 'scripts/gmaps_2.js' %}"></script>
-<script defer src="{% static 'scripts/survey_2.js' %}"></script>
+<script defer src="{% static 'scripts/gmaps.js' %}"></script>
+<script defer src="{% static 'scripts/survey.js' %}"></script>
 {% endcompress %}
 
 {% endblock page_scripts %}
